+++ rd-api-client/src/main/java/org/rundeck/client/api/model/ScmImportItem.java
@@ -17,6 +17,7 @@
 package org.rundeck.client.api.model;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+import org.rundeck.toolbelt.Formatable;
 
 import java.util.HashMap;
 import java.util.List;