+++ rd-api-client/src/main/java/org/rundeck/client/api/model/ScmInputField.java
@@ -16,6 +16,7 @@
 
 package org.rundeck.client.api.model;
 
+import org.rundeck.toolbelt.Formatable;
 
 import java.util.HashMap;
 import java.util.List;