+++ src/main/java/org/rundeck/client/tool/commands/Nodes.java
@@ -18,9 +18,9 @@ package org.rundeck.client.tool.commands;
 
 import com.lexicalscope.jewel.cli.CommandLineInterface;
 import com.lexicalscope.jewel.cli.Unparsed;
-import com.simplifyops.toolbelt.Command;
-import com.simplifyops.toolbelt.CommandOutput;
-import com.simplifyops.toolbelt.InputError;
+import org.rundeck.toolbelt.Command;
+import org.rundeck.toolbelt.CommandOutput;
+import org.rundeck.toolbelt.InputError;
 import org.rundeck.client.api.model.ProjectNode;
 import org.rundeck.client.tool.RdApp;
 import org.rundeck.client.tool.options.NodeFilterOptions;