+++ src/main/java/org/rundeck/client/tool/commands/projects/Readme.java
@@ -17,9 +17,9 @@
 package org.rundeck.client.tool.commands.projects;
 
 import com.lexicalscope.jewel.cli.Option;
-import com.simplifyops.toolbelt.Command;
-import com.simplifyops.toolbelt.CommandOutput;
-import com.simplifyops.toolbelt.InputError;
+import org.rundeck.toolbelt.Command;
+import org.rundeck.toolbelt.CommandOutput;
+import org.rundeck.toolbelt.InputError;
 import okhttp3.MediaType;
 import okhttp3.RequestBody;
 import org.rundeck.client.api.ReadmeFile;