+++ src/main/java/org/rundeck/client/tool/commands/system/Mode.java
@@ -2,9 +2,9 @@ package org.rundeck.client.tool.commands.system;
 
 import com.lexicalscope.jewel.cli.CommandLineInterface;
 import com.lexicalscope.jewel.cli.Option;
-import com.simplifyops.toolbelt.Command;
-import com.simplifyops.toolbelt.CommandOutput;
-import com.simplifyops.toolbelt.InputError;
+import org.rundeck.toolbelt.Command;
+import org.rundeck.toolbelt.CommandOutput;
+import org.rundeck.toolbelt.InputError;
 import org.rundeck.client.api.RundeckApi;
 import org.rundeck.client.api.model.ExecutionMode;
 import org.rundeck.client.api.model.SystemInfo;