+++ src/test/groovy/org/rundeck/client/tool/commands/projects/ACLsSpec.groovy
@@ -1,6 +1,6 @@
 package org.rundeck.client.tool.commands.projects
 
-import com.simplifyops.toolbelt.CommandOutput
+import org.rundeck.toolbelt.CommandOutput
 import okhttp3.ResponseBody
 import org.rundeck.client.api.RequestFailed
 import org.rundeck.client.api.RundeckApi