+++ src/test/groovy/org/rundeck/client/tool/commands/projects/SCMSpec.groovy
@@ -16,7 +16,7 @@
 
 package org.rundeck.client.tool.commands.projects
 
-import com.simplifyops.toolbelt.CommandOutput
+import org.rundeck.toolbelt.CommandOutput
 import okhttp3.ResponseBody
 import org.rundeck.client.api.RundeckApi
 import org.rundeck.client.api.model.ScmActionInputsResult